+++ lib/liquid_engine.dart
@@ -1,7 +1,6 @@
 library liquid_engine;
 
 export 'src/block.dart' show Block, SimpleBlockFactory, BlockParser;
-export 'src/context.dart'
-    show Context, Filter, Module, ParseContext, RenderContext;
+export 'src/context.dart' show Context, Filter, Module, ParseContext, RenderContext;
 export 'src/model.dart' show Source, Root;
 export 'src/template.dart' show Template;