+++ lib/src/buildin_tags/for.dart
@@ -17,17 +17,18 @@ class For extends Block {
   For(this.to, this.from, this.innerChildren, this.elseChildren) : super([]);
 
   @override
-  Iterable<String> render(RenderContext context) {
-    var collection = List.from(from.evaluate(context) ?? []);
+  Stream<String> render(RenderContext context) async* {
+    var collection = List.from(await from.evaluate(context) ?? []);
 
     if (collection.isEmpty) {
-      return super.renderTags(context, elseChildren);
+      yield* super.renderTags(context, elseChildren);
+      return;
     }
 
     var parentLoop = context.variables['forloop'];
 
     var index = 0;
-    return collection.map<Iterable<String>>((item) {
+    for (final item in collection) {
       final innerContext = context.push({
         to: item,
         'forloop': {
@@ -47,15 +48,24 @@ class For extends Block {
         }
       });
       index++;
-      return super.renderTags(innerContext, innerChildren);
-    }).flatten();
+      yield* super.renderTags(innerContext, innerChildren);
+    }
   }
 
   static final BlockParserFactory factory = () => _ForBlockParser();
+
+  @override
+  String toString() {
+    var innerC = innerChildren.toString();
+    if (innerC.length > 40) innerC = innerC.substring(0, 30) + '...' + innerC.substring(innerC.length - 7);
+    var elseC = elseChildren.toString();
+    if (elseC.length > 40) elseC = elseC.substring(0, 30) + '...' + elseC.substring(elseC.length - 7);
+    return 'For{from: $from, to: $to, innerChildren: $innerC, elseChildren: $elseC}';
+  }
 }
 
 class _ForBlockParser extends BlockParser {
-  List<Tag> innerChildren;
+  List<Tag>? innerChildren;
 
   @override
   void start(context, args) {
@@ -73,13 +83,11 @@ class _ForBlockParser extends BlockParser {
     parser.expect(types: [TokenType.identifier], value: 'in');
 
     parser.moveNext();
-    return For(to, parser.parseFilterExpression(), innerChildren ?? children,
-        innerChildren != null ? children : []);
+    return For(to, parser.parseFilterExpression(), innerChildren ?? children, innerChildren != null ? children : []);
   }
 
   @override
-  void unexpectedTag(
-      Parser parser, Token start, List<Token> args, List<Tag> childrenSoFar) {
+  void unexpectedTag(Parser parser, Token start, List<Token> args, List<Tag> childrenSoFar) {
     if (start.value == 'else' || start.value == 'empty') {
       if (innerChildren != null) {
         throw ParseException('Only one {% else %} is allowed in a {% for %}');
@@ -87,8 +95,7 @@ class _ForBlockParser extends BlockParser {
       innerChildren = List.from(childrenSoFar);
       childrenSoFar.clear();
     } else {
-      throw ParseException.unexpected(start,
-          expected: '{% else %} or {% endfor %}');
+      throw ParseException.unexpected(start, expected: '{% else %} or {% endfor %}');
     }
   }
 }