+++ lib/src/buildin_tags/if.dart
@@ -13,14 +13,14 @@ class If extends Block {
   If(this.conditions) : super([]);
 
   @override
-  Iterable<String> render(RenderContext context) {
+  Stream<String> render(RenderContext context) async* {
     for (final condition in conditions) {
-      final test = condition.key.evaluate(context);
+      final test = await condition.key.evaluate(context);
       if (test) {
-        return renderTags(context, condition.value);
+        yield* renderTags(context, condition.value);
+        return;
       }
     }
-    return Iterable.empty();
   }
 
   static final BlockParserFactory factory = () => _IfBlockParser();
@@ -28,20 +28,19 @@ class If extends Block {
 }
 
 class _IfBlockParser extends BlockParser {
-  List<MapEntry<Expression, List<Tag>>> conditions;
-  Expression lastExpression;
+  final List<MapEntry<Expression, List<Tag>>> conditions = [];
+  Expression? lastExpression;
 
   @override
   void start(context, args) {
     super.start(context, args);
-    conditions = [];
+    conditions.clear();
     lastExpression = TagParser.from(args).parseBooleanExpression();
   }
 
   @override
   Block create(List<Token> tokens, List<Tag> children) {
-    conditions.add(MapEntry(
-        lastExpression ?? ConstantExpression(true), List.from(children)));
+    conditions.add(MapEntry(lastExpression ?? ConstantExpression(true), List.from(children)));
     return createFromConditions(conditions);
   }
 
@@ -50,26 +49,23 @@ class _IfBlockParser extends BlockParser {
   }
 
   @override
-  void unexpectedTag(
-      Parser parser, Token start, List<Token> args, List<Tag> childrenSoFar) {
+  void unexpectedTag(Parser parser, Token start, List<Token> args, List<Tag> childrenSoFar) {
     if (start.value == 'elseif') {
       if (lastExpression == null) {
-        throw ParseException.unexpected(start,
-            expected: '{% elseif %} must preced {% else %}');
+        throw ParseException.unexpected(start, expected: '{% elseif %} must preced {% else %}');
       }
-      conditions.add(MapEntry(lastExpression, List.from(childrenSoFar)));
+      conditions.add(MapEntry(lastExpression!, List.from(childrenSoFar)));
       childrenSoFar.clear();
       lastExpression = TagParser.from(args).parseBooleanExpression();
     } else if (start.value == 'else') {
       if (lastExpression == null) {
         throw ParseException.unexpected(start, expected: '{% endif %}');
       }
-      conditions.add(MapEntry(lastExpression, List.from(childrenSoFar)));
+      conditions.add(MapEntry(lastExpression!, List.from(childrenSoFar)));
       childrenSoFar.clear();
       lastExpression = null;
     } else {
-      throw ParseException.unexpected(start,
-          expected: '{% elseif %} or {% else %} or {% endif %}');
+      throw ParseException.unexpected(start, expected: '{% elseif %} or {% else %} or {% endif %}');
     }
   }
 }