+++ lib/src/buildin_tags/include.dart
@@ -12,23 +12,25 @@ class Include extends Block {
   final bool clearVariables;
   final DocumentFuture childBuilder;
 
-  Include._(this.assignments, this.clearVariables, this.childBuilder)
-      : super([]);
+  Include._(this.assignments, this.clearVariables, this.childBuilder) : super([]);
 
   @override
-  Iterable<String> render(RenderContext context) {
+  Stream<String> render(RenderContext context) async* {
     var innerContext = context;
     if (clearVariables) {
       innerContext = innerContext.clone();
       innerContext.variables.clear();
     }
-    innerContext = innerContext.push(Map.fromIterable(
-      assignments,
-      key: (a) => a.to,
-      value: (a) => a.from.evaluate(context),
-    ));
 
-    return childBuilder.resolve(context).render(innerContext);
+    /// [old]
+    // innerContext = innerContext.push(Map.fromIterable(
+    //   assignments,
+    //   key: (a) => a.to,
+    //   value: (a) => a.from.evaluate(context),
+    // ));
+    innerContext = innerContext.push({for (var a in assignments) a.to: a.from.evaluate(context)});
+
+    yield* (await childBuilder.resolve(context)).render(innerContext);
   }
 
   static BlockParserFactory factory = () => _IncludeBlockParser();
@@ -51,10 +53,9 @@ class _IncludeBlockParser extends BlockParser {
     final childBuilder = parser.parseDocumentReference(context);
 
     final assignments = <_Assign>[];
-    if (parser.current != null && parser.current.value == 'with') {
+    if (parser.current.value == 'with') {
       parser.moveNext();
-      while (parser.current.type == TokenType.identifier &&
-          parser.current.value != 'only') {
+      while (parser.current.type == TokenType.identifier && parser.current.value != 'only') {
         parser.expect(types: [TokenType.identifier]);
         final to = parser.current;
 
@@ -67,13 +68,11 @@ class _IncludeBlockParser extends BlockParser {
       }
     }
 
-    final clearVariable =
-        parser.current != null && parser.current.value == 'only';
+    final clearVariable = parser.current.value == 'only';
 
     return Include._(assignments, clearVariable, childBuilder);
   }
 
   @override
-  void unexpectedTag(
-      Parser parser, Token start, List<Token> args, List<Tag> childrenSoFar) {}
+  void unexpectedTag(Parser parser, Token start, List<Token> args, List<Tag> childrenSoFar) {}
 }