+++ lib/src/context.dart
@@ -2,14 +2,14 @@ import './builtins.dart';
 import 'block.dart';
 import 'tag.dart';
 
-typedef dynamic Filter(dynamic input, List<dynamic> args);
+typedef Filter = dynamic Function(dynamic input, List<dynamic> args);
 
 abstract class RenderContext {
-  RenderContext get parent;
+  RenderContext? get parent;
 
   RenderContext get root;
 
-  Map<String, Iterable<String>> blocks;
+  Map<String, String> get blocks;
 
   Map<String, dynamic> get variables;
 
@@ -36,7 +36,7 @@ abstract class Module {
 
 class Context implements RenderContext, ParseContext {
   @override
-  Map<String, Iterable<String>> blocks = {};
+  Map<String, String> blocks = {};
 
   @override
   Map<String, BlockParserFactory> tags = {};
@@ -49,53 +49,48 @@ class Context implements RenderContext, ParseContext {
 
   Map<String, Module> modules = {'builtins': BuiltinsModule()};
 
-  var _tagStates = {};
+  Map<Tag, Map<String, dynamic>> tagStates = {};
 
   @override
-  Map<String, dynamic> getTagState(Tag tag) => parent == null
-      ? _tagStates.putIfAbsent(tag, () => {})
-      : root.getTagState(tag);
+  Map<String, dynamic> getTagState(Tag tag) => parent == null ? tagStates.putIfAbsent(tag, () => {}) : root.getTagState(tag);
 
   @override
-  Context parent;
+  final Context? parent;
 
   @override
-  RenderContext get root => parent == null ? this : parent.root;
+  RenderContext get root => parent == null ? this : parent!.root;
 
-  Context._();
+  Context._({Context? parent}) : parent = parent;
 
   @override
   void registerModule(String load) {
-    modules[load].register(this);
+    modules[load]!.register(this);
   }
 
   factory Context.create() {
     final context = Context._();
 
-    context.modules['builtins'].register(context);
+    context.modules['builtins']!.register(context);
 
     return context;
   }
 
   @override
-  Context push(Map<String, dynamic> variables) {
-    final context = Context._();
+  Context push(Map<String, dynamic> variables, {Context? root}) {
+    final context = Context._(parent: root ?? this);
     context.blocks = Map.from(blocks);
     context.tags = Map.from(tags);
     context.filters = Map.from(filters);
     context.variables = Map.from(this.variables);
     context.variables.addAll(variables);
-    context.parent = this;
     return context;
   }
 
   @override
-  Context clone() => push({});
+  Context clone({Context? root}) => push({}, root: root);
 
   @override
   Context cloneAsRoot() {
-    final context = clone();
-    context.parent = null;
-    return context;
+    return clone(root: null);
   }
 }