+++ lib/src/document.dart
@@ -15,28 +15,34 @@ class DocumentFuture {
   final ParseContext context;
   final Expression path;
 
+  Document? _document;
+
   DocumentFuture(this.root, this.context, this.path);
 
-  Document resolve(RenderContext context) {
-    var path = this.path.evaluate(context);
+  Future<Document> resolve(RenderContext context) async {
+    return _document ?? (_document = await _resolve(context));
+  }
+
+  Future<Document> _resolve(RenderContext context) async {
+    var path = await this.path.evaluate(context);
     if (path is Template) {
       return path.document;
     }
     if (path is Document) {
       return path;
     }
-    return Template.parse(this.context, root.resolve(path)).document;
+    return Template.parse(this.context, await root.resolve(path)).document;
   }
 }
 
 class Document extends Block {
-  DocumentFuture base;
+  DocumentFuture? base;
   List<String> loads;
 
   Document(this.base, this.loads, List<Tag> children) : super(children);
 
   @override
-  Iterable<String> render(RenderContext initialContext) {
+  Stream<String> render(RenderContext initialContext) async* {
     var context = initialContext.cloneAsRoot();
 
     for (final load in loads) {
@@ -44,24 +50,25 @@ class Document extends Block {
     }
 
     if (base == null) {
-      return super.render(context);
+      yield* super.render(context);
+      return;
     }
 
     var baseContext = initialContext.cloneAsRoot();
     for (final tag in children) {
       if (tag is NamedBlock) {
-        baseContext.blocks[tag.name] = tag.render(context);
+        baseContext.blocks[tag.name] = await tag.render(context).join();
       } else {
         tag.render(context);
       }
     }
-    return base.resolve(context).render(baseContext);
+    yield* (await base!.resolve(context)).render(baseContext);
   }
 }
 
 class DocumentParser extends BlockParser {
   @override
-  bool approveTag(Token start, List<Tag> childrenSoFar, Token asToken) {
+  bool approveTag(Token start, List<Tag> childrenSoFar, Token? asToken) {
     if (start.value == 'extends') {
       return childrenSoFar.isEmpty;
     }
@@ -77,7 +84,7 @@ class DocumentParser extends BlockParser {
   @override
   Block create(List<Token> tokens, List<Tag> children) {
     var start = 0;
-    DocumentFuture base;
+    DocumentFuture? base;
     final loads = <String>[];
     if (children.isNotEmpty) {
       if (children.length > start && children[start] is Extends) {
@@ -94,8 +101,7 @@ class DocumentParser extends BlockParser {
   }
 
   @override
-  void unexpectedTag(
-      Parser parser, Token start, List<Token> args, List<Tag> childrenSoFar) {
+  void unexpectedTag(Parser parser, Token start, List<Token> args, List<Tag> childrenSoFar) {
     throw ParseException.unexpected(start);
   }
 }