+++ lib/src/errors.dart
@@ -8,6 +8,9 @@ class ParseException extends Error {
   @override
   String toString() => 'Parse Exception: $message';
 
-  ParseException.unexpected(Token token, {dynamic expected})
-      : this('Unexpected Token: $token expected $expected');
+  ParseException.unexpected(Token token, {dynamic expected}) : this('Unexpected Token: $token expected $expected');
+
+  ParseException.error() : this('Unexpected Error');
+
+  ParseException.missingRoot() : this('To reference external document you must specify a root directory.');
 }