+++ lib/src/exception/tag_render_exception.dart
@@ -1,0 +1,16 @@
+import '../tag.dart';
+
+class TagRenderException implements Exception {
+  final error;
+  final stacktrace;
+  final Tag tag;
+
+  TagRenderException(this.error, this.stacktrace, this.tag);
+
+  @override
+  String toString() {
+    //return super.toString();
+    var result = 'TagRenderException at tag $tag:\n  $error\nStacktrace was at:\n$stacktrace';
+    return result;
+  }
+}