+++ lib/src/expressions.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import './errors.dart';
 import 'block.dart';
 import 'context.dart';
@@ -5,10 +7,14 @@ import 'model.dart';
 import 'tag.dart';
 
 abstract class Expression {
-  dynamic evaluate(RenderContext context);
+  Future<dynamic> evaluate(RenderContext context);
+  @override
+  String toString() {
+    return 'Expression{}';
+  }
 }
 
-typedef dynamic Operation(dynamic a, dynamic b);
+typedef Operation = dynamic Function(dynamic a, dynamic b);
 
 class BooleanCastExpression implements Expression {
   final Expression input;
@@ -16,7 +22,7 @@ class BooleanCastExpression implements Expression {
   BooleanCastExpression(this.input);
 
   @override
-  evaluate(RenderContext context) {
+  Future<bool> evaluate(RenderContext context) async {
     bool _bool(dynamic a) {
       if (a == null) {
         return false;
@@ -34,7 +40,12 @@ class BooleanCastExpression implements Expression {
       return true;
     }
 
-    return _bool(input.evaluate(context));
+    return _bool(await input.evaluate(context));
+  }
+
+  @override
+  String toString() {
+    return 'BooleanCastExpression{input: $input}';
   }
 }
 
@@ -42,7 +53,11 @@ class NotExpression extends BooleanCastExpression {
   NotExpression(Expression input) : super(input);
 
   @override
-  evaluate(RenderContext context) => !super.evaluate(context);
+  Future<bool> evaluate(RenderContext context) async => !(await super.evaluate(context));
+  @override
+  String toString() {
+    return 'NotExpression{}';
+  }
 }
 
 class BinaryOperation implements Expression {
@@ -52,8 +67,12 @@ class BinaryOperation implements Expression {
 
   BinaryOperation(this.operation, this.left, this.right);
 
-  dynamic evaluate(RenderContext context) =>
-      operation(left.evaluate(context), right.evaluate(context));
+  @override
+  Future<dynamic> evaluate(RenderContext context) async => operation(await left.evaluate(context), await right.evaluate(context));
+  @override
+  String toString() {
+    return 'BinaryOperation{operation: $operation, left: $left, right: $right}';
+  }
 }
 
 class ConstantExpression implements Expression {
@@ -62,10 +81,8 @@ class ConstantExpression implements Expression {
   ConstantExpression(this.value);
 
   factory ConstantExpression.fromToken(Token token) {
-    if (token.type == TokenType.single_string ||
-        token.type == TokenType.double_string) {
-      return ConstantExpression(
-          token.value.substring(1, token.value.length - 1));
+    if (token.type == TokenType.single_string || token.type == TokenType.double_string) {
+      return ConstantExpression(token.value.substring(1, token.value.length - 1));
     } else if (token.type == TokenType.number) {
       return ConstantExpression(num.parse(token.value));
     } else {
@@ -74,7 +91,11 @@ class ConstantExpression implements Expression {
   }
 
   @override
-  evaluate(RenderContext context) => value;
+  Future<dynamic> evaluate(RenderContext context) async => value;
+  @override
+  String toString() {
+    return 'ConstantExpression{value: $value}';
+  }
 }
 
 class LookupExpression implements Expression {
@@ -83,8 +104,13 @@ class LookupExpression implements Expression {
   LookupExpression(this.name);
 
   @override
-  evaluate(RenderContext context) {
+  Future<dynamic> evaluate(RenderContext context) async {
     return context.variables[name.value];
+  }
+
+  @override
+  String toString() {
+    return 'LookupExpression{name: $name}';
   }
 }
 
@@ -95,8 +121,8 @@ class MemberExpression implements Expression {
   MemberExpression(this.base, this.member);
 
   @override
-  evaluate(RenderContext context) {
-    final base = this.base.evaluate(context);
+  Future<dynamic> evaluate(RenderContext context) async {
+    final base = await this.base.evaluate(context);
     if (base == null) {
       return null;
     }
@@ -106,8 +132,14 @@ class MemberExpression implements Expression {
     if (base is List) {
       return base[int.parse(member.value) % base.length];
     } else {
+      print('Warning: $this only supports bases of type List or Map, found ${base.runtimeType} instead, returning null');
       return null;
     }
+  }
+
+  @override
+  String toString() {
+    return 'MemberExpression{base: $base, member: $member}';
   }
 }
 
@@ -118,13 +150,20 @@ class FilterExpression implements Expression {
 
   FilterExpression(this.input, this.name, this.arguments);
 
-  dynamic evaluate(RenderContext context) {
-    var output = input.evaluate(context);
+  @override
+  Future<dynamic> evaluate(RenderContext context) async {
+    var output = await input.evaluate(context);
     var filter = context.filters[name.value];
     if (filter == null) {
       throw ParseException.unexpected(name);
     }
-    return filter(output, arguments.map((e) => e.evaluate(context)).toList());
+    return filter(output, await Future.wait(arguments.map((e) => _evaluateToFuture(e, context))));
+  }
+
+  Future<dynamic> _evaluateToFuture(Expression e, RenderContext context) async => await e.evaluate(context);
+  @override
+  String toString() {
+    return 'FilterExpression{input: $input, name: $name, arguments: $arguments}';
   }
 }
 
@@ -136,17 +175,26 @@ class BlockExpression implements Expression {
   BlockExpression.fromTags(List<Tag> tags) : this(Block(tags));
 
   @override
-  String evaluate(RenderContext context) => block.render(context).join('');
+  Future<String> evaluate(RenderContext context) => block.render(context).join('');
+  @override
+  String toString() {
+    return 'BlockExpression{block: $block}';
+  }
 }
 
 class ExpressionTag implements Tag {
   Expression expression;
 
   @override
-  Iterable<String> render(RenderContext context) sync* {
-    var output = expression.evaluate(context);
+  Stream<String> render(RenderContext context) async* {
+    var output = await expression.evaluate(context);
     yield output != null ? output.toString() : '';
   }
 
   ExpressionTag(this.expression);
+
+  @override
+  String toString() {
+    return 'ExpressionTag{expression: $expression}';
+  }
 }