+++ lib/src/model.dart
@@ -1,21 +1,13 @@
-import 'dart:io';
+export 'model_web.dart' if (dart.library.io) 'model_io.dart';
 
-class Root {
-  final Uri path;
-
-  Root(this.path);
-
-  Source resolve(String relPath) {
-    final file = path.resolve(relPath);
-    final content = File.fromUri(file).readAsStringSync();
-    return Source(file, content, this);
-  }
+abstract class Root {
+  Future<Source> resolve(String relPath);
 }
 
 class Source {
-  final Uri file;
+  final Uri? file;
   final String content;
-  final Root root;
+  final Root? root;
 
   Source(this.file, this.content, this.root);
 
@@ -45,16 +37,17 @@ enum TokenType {
   var_start,
   var_end,
   markup,
+  eof,
 }
 
 class Token {
-  final TokenType type;
+  final TokenType? type;
   final String value;
-  final Source source;
-  final int line;
-  final int column;
+  final Source? source;
+  final int? line;
+  final int? column;
 
-  static Token eof = Token(null, '<EOF>');
+  static Token eof = Token(TokenType.eof, '<EOF>');
 
   Token(this.type, this.value, {this.source, this.line, this.column});