+++ lib/src/parser/parser.dart
@@ -1,3 +1,5 @@
+import '../exception/parse_block_exception.dart';
+
 import '../block.dart';
 import '../context.dart';
 import '../document.dart';
@@ -13,15 +15,13 @@ class Parser {
   final Iterator<Token> tokens;
   final ParseContext context;
 
-  Parser(this.context, this.source)
-      : this.tokens = Lexer().tokenize(source).iterator;
+  Parser(this.context, this.source) : tokens = Lexer().tokenize(source).iterator;
 
   Document parse() {
     return parseBlock(DocumentParser(), [], '<doc>', null) as Document;
   }
 
-  Block parseBlock(
-      BlockParser builder, List<Token> args, String start, Token asTarget) {
+  Block parseBlock(BlockParser builder, List<Token> args, String start, Token? asTarget) {
     var innerChildren = <Tag>[];
     builder.start(context, args);
     if (builder.hasEndTag) {
@@ -31,11 +31,15 @@ class Parser {
         builder,
       );
     }
-    var block = builder.create(args, innerChildren);
-    if (asTarget != null) {
-      block = AsBlock(asTarget.value, [block]);
+    try {
+      var block = builder.create(args, innerChildren);
+      if (asTarget != null) {
+        block = AsBlock(asTarget.value, [block]);
+      }
+      return block;
+    } catch (error) {
+      throw ParseBlockException(error.toString(), start, args);
     }
-    return block;
   }
 
   void parseBlockChildren(
@@ -52,8 +56,7 @@ class Parser {
           expect(types: [TokenType.identifier]);
           final start = tokens.current;
           final args = <Token>[];
-          while (
-              tokens.moveNext() && tokens.current.type != TokenType.tag_end) {
+          while (tokens.moveNext() && tokens.current.type != TokenType.tag_end) {
             args.add(tokens.current);
           }
 
@@ -61,16 +64,14 @@ class Parser {
             return;
           }
 
-          Token asTarget;
-          if (args.length >= 2 &&
-              args[args.length - 1].type == TokenType.identifier &&
-              args[args.length - 2].value == 'as') {
+          Token? asTarget;
+          if (args.length >= 2 && args[args.length - 1].type == TokenType.identifier && args[args.length - 2].value == 'as') {
             asTarget = args.last;
             args.length = args.length - 2;
           }
 
           if (context.tags.containsKey(start.value)) {
-            var builder = context.tags[start.value]();
+            var builder = context.tags[start.value]!();
             if (!parent.approveTag(start, children, asTarget)) {
               throw ParseException.unexpected(start);
             }
@@ -97,11 +98,11 @@ class Parser {
 
     var exp = TagParser.fromIterator(tokens).parseFilterExpression();
 
-    expect(value: '}}');
+    expect(types: [TokenType.var_end]);
     return ExpressionTag(exp);
   }
 
-  void expect({List<TokenType> types, String value, Token token}) {
+  void expect({List<TokenType>? types, String? value, Token? token}) {
     token ??= tokens.current;
     if (types != null && !types.contains(token.type)) {
       throw ParseException.unexpected(token, expected: 'one of $types');