+++ lib/src/template.dart
@@ -4,20 +4,20 @@ import 'model.dart';
 import 'parser/parser.dart';
 
 extension JsSplit on String {
-  Iterable<String> inclusiveSplit(RegExp pattern) sync* {
+  Iterable<String?> inclusiveSplit(RegExp pattern) sync* {
     final matches = pattern.allMatches(this);
     var pos = 0;
 
     for (final match in matches) {
       if (match.start > pos) {
-        yield this.substring(pos, match.start);
+        yield substring(pos, match.start);
       }
-      yield match.group(0);
+      yield match.group(0)!;
       pos = match.end;
     }
 
-    if (pos < this.length - 1) {
-      yield this.substring(pos);
+    if (pos < length - 1) {
+      yield substring(pos);
     }
   }
 }
@@ -28,9 +28,13 @@ class Template {
 
   Template(this.source, this.document);
 
-  factory Template.parse(ParseContext context, Source source) =>
-      Template(source, Parser(context, source).parse());
+  factory Template.parse(ParseContext context, Source source) => Template(source, Parser(context, source).parse());
 
-  String render(Context context) =>
-      (StringBuffer()..writeAll(document.render(context))).toString();
+  Future<String> render(Context context) async {
+    final buffer = StringBuffer();
+    await for (final chunk in document.render(context)) {
+      buffer.write(chunk);
+    }
+    return buffer.toString();
+  }
 }