+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## 0.2.2
+
+- support for regroup tag
+- more verbose error messages
+
+## 0.2.1
+
+- support for `true` and `false`
+- support for whitespace control
+- async rendering 
+- web support
+- a few minor bug fixes
+
+## 0.2.0
+
+- null safety
+- a few minor bug fixes
+
 ## 0.1.0
 
 - Initial version